\name{bellevue.wwtp.ponds}

\alias{bellevue.wwtp.ponds}

\docType{data}

\title{Bellevue Waste Water Treatment Plant Ponds}

\description{
The location of the Bellevue Waste Water Treatment Plant ponds.
}

\usage{bellevue.wwtp.ponds}

\format{
An object of \code{SpatialPolygons} class containing 1 \code{Polygons}.
Geographic coordinates are in units of meters, in conformance with the North American Datum of 1983 (NAD 83), and placed in the
Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}

\source{
Idaho Department of Water Resources, accessed on December 11, 2014
}

\examples{
plot(bellevue.wwtp.ponds)
}

\keyword{datasets}
