\name{bypass.canal}

\alias{bypass.canal}

\docType{data}

\title{Bypass Canal}

\description{
The location of the Bypass Canal.
}

\usage{bypass.canal}

\format{
An object of \code{SpatialLines} class containing 4 \code{Lines}.
Geographic coordinates are in units of meters, in conformance with the North American Datum of 1983 (NAD 83), and placed in the
Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}

\source{
Idaho Department of Water Resources, accessed on January 15, 2015
}

\examples{
plot(bypass.canal)
}

\keyword{datasets}
