% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canal.seep}
\alias{canal.seep}
\title{Canal Seepage}
\format{
An object of class data.frame with 19 records and the following variables:
  \describe{
    \item{EntityName}{name of the irrigation entity served by the canal system.}
    \item{SeepFrac}{estimated canal seepage as a fraction of diversions.}
  }
}
\source{
Idaho Department of Water Resources, accessed on November 4, 2015
}
\usage{
canal.seep
}
\description{
Canal seepage as a fraction of diversions for irrigation entities in the Wood River Valley.
}
\examples{
str(canal.seep)

d <- canal.seep[order(canal.seep$SeepFrac, decreasing=TRUE), ]
par(mar = c(4.1, 8.1, 0.1, 0.6))
barplot(d$SeepFrac, names.arg = d$EntityName, horiz = TRUE, cex.names = 0.7,
        cex.axis = 0.7, cex.lab = 0.7, las = 1, xlab = "Seepage fraction")

graphics.off()

}
\seealso{
\code{\link{canals}}
}
\keyword{datasets}
