% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{div.gw}
\alias{div.gw}
\title{Groundwater Diversions}
\format{
An object of class data.frame with 7,292 records and the following variables:
  \describe{
    \item{YearMonth}{year and month during which diversions were recorded,
      with a required date format of YYYYMM.}
    \item{Diversion}{name of the well}
    \item{Reach}{name of the river subreach into which the well water is discharged;
      only applicable to exchange wells.}
    \item{BigReach}{name of the river reach into which the well water is discharged;
      only applicable to exchange wells.}
    \item{EntityName}{name of the irrigation entity which the well supplies water.}
    \item{WMISNumber}{well number in the Idaho Department of Water Resources (IDWR)
      Water Measurement Information System.}
    \item{GWDiv}{volume of water diverted during the month, in cubic meters.}
  }
}
\source{
Idaho Department of Water Resources (IDWR), accessed on December 11, 2014;
  compiled data records from Water District 37 and 37M, City of Ketchum,
  Sun Valley Water and Sewer District, City of Hailey, and City of Bellevue.
}
\usage{
div.gw
}
\description{
Groundwater diversions recorded by Water District 37 or municipal water providers.
Groundwater is diverted from the aquifer by means of either pumping wells or
flowing-artesian wells.
}
\examples{
str(div.gw)

}
\keyword{datasets}
