% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{div.ww}
\alias{div.ww}
\title{Wastewater Treatment Plant Diversions}
\format{An object of class data.frame with 1,182 records and the following variables:
  \describe{
    \item{YearMonth}{year and month during which diversions were recorded,
      with a required date format of YYYYMM.}
    \item{Return}{name of the wastewater treatment plant.}
    \item{Reach}{name of the river subreach to which treated effluent is discharged;
      only applicable to wastewater treatment plants that discharge to the river.}
    \item{BigReach}{name of the river reach to which treated effluent is discharged;
      only applicable to wastewater treatment plants that discharge to the river.}
    \item{EntityName}{name of the irrigation entity served by the wastewater treatment plant.}
    \item{WWDiv}{volume of wastewater discharged during the month, in cubic meters.}
  }}
\source{
Idaho Department of Water Resources and U.S. Geological Survey,
  accessed on August 11, 2014; compiled data records from the
  U.S. Environmental Protection Agency for plants that discharge to the river,
  and from records of the Idaho Department of Environmental Quality for
  plants that discharge to land application.
}
\usage{
div.ww
}
\description{
Discharge from wastewater treatment plants.
}
\examples{
str(div.ww)

}
\keyword{datasets}

