% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entity.components}
\alias{entity.components}
\title{Irrigation Entity Components}
\format{
An object of list class with components of SpatialPolygonsDataFrame-class.
  There are a total of 192 components, one for each month in the 1995--2010 time period.
  Linked data.frame objects have the following variables:
  \describe{
    \item{EntitySrce}{concatenation of the \code{EntityName} and \code{Source} character strings.}
    \item{mean.et}{mean evapotranspiration (ET) on irrigated and semi-irrigated lands in meters.}
    \item{area}{area of irrigated and semi-irrigated lands in square meters.}
    \item{PrecipZone}{name of the precipitation zone;
      see \code{\link{precip.zones}} dataset for details.}
    \item{et.vol}{volume of ET on irrigated and semi-irrigated lands in cubic meters.}
    \item{precip.vol}{volume of precipitation on irrigated and semi-irrigated lands in cubic meters.}
    \item{cir.vol}{volume of crop irrigation requirement in cubic meters (ET minus precipitation).}
    \item{EntityName}{name of the irrigation entity.}
    \item{Source}{water source, either \dQuote{Mixed} for a mixture of surface water and groundwater,
      \dQuote{SW Only} for surface water only, or \dQuote{GW Only} for groundwater only.}
  }
}
\source{
Calculated from the \code{\link{irr.entities}}, \code{\link{wetlands}},
  \code{\link{public.parcels}}, \code{\link{irr.lands.year}}, \code{\link{et}}, and
  \code{\link{precipitation}} datasets;
  see the \file{package-datasets} vignette for the \R code used in this calculation.
}
\usage{
entity.components
}
\description{
Irrigation entities and their components in the Wood River Valley and surrounding areas.
An irrigation entity is defined as an area served by a group of surface-water and/or
groundwater diversion(s).
}
\examples{
names(entity.components)
sp::plot(entity.components[["199506"]])
print(entity.components[["199506"]])

}
\keyword{datasets}
