% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{et.method}
\alias{et.method}
\title{Method Used to Calculate Evapotranspiration}
\format{
An object of data.frame class with 122 records with the following variables:
  \describe{
    \item{YearMonth}{year and month during which the method was applied,
      with a required date format of YYYYMM.}
    \item{ETMethod}{Identifier that indicates the method used to estimate ET values.
      Identifiers include either
        \dQuote{Allen-Robison}, the Allen and Robison method (Allen and Robison, 2007);
        \dQuote{METRIC}, the Mapping Evapotranspiration at high Resolution and with
          Internalized Calibration (METRIC) model (Allen and others, 2010a);
        \dQuote{NDVI}, the Normalized Difference Vegetation Index (NDVI) method
          (Allen and others, 2010b);
        \dQuote{Interpolation}, interpolation from known ET data; or
        \dQuote{METRIC-NDVI}, a combination of METRIC and NDVI methods.}
  }
}
\source{
Idaho Department of Water Resources, accessed on April 27, 2015
}
\usage{
et.method
}
\description{
Methods used to estimate monthly distributions of evapotranspiration (ET) rate.
}
\examples{
str(et.method)

}
\references{
Allen, R., and Robison, C.W., 2007, Evapotranspiration and
  consumptive water requirements for Idaho, University of Idaho, Kimberly, Idaho.

  Allen, R., Tasumi, M., Trezza, R., and Kjaersgaard, J., 2010a,
    METRIC mapping evapotranspiration at high resolution applications manual for
    Landsat satellite imagery version 2.07, University of Idaho, Kimberly, ID.

  Allen, R., Robison, C.W., Garcia, M., Trezza, R., Tasumi, M., and Kjaersgaard, J., 2010b,
    ETrF vs NDVI relationships for southern Idaho for rapid estimation of evapotranspiration,
    University of Idaho, Kimberly, ID.

  ET Idaho: \url{http://data.kimberly.uidaho.edu/ETIdaho/}
}
\keyword{datasets}
