% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hill.shading}
\alias{hill.shading}
\title{Land Surface Hill Shading}
\format{An object of \code{RasterLayer} class.
  Each cell on the surface grid represents the hill shade.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
  The spatial grid is composed of 3,108 rows and 2,360 columns,
  and has cell sizes that are constant at 20 meters by 20 meters.}
\source{
Calculated from the slope and aspect of the \code{\link{land.surface}} dataset
  using the \code{terrain} and \code{hillShade} functions;
  see the appendix C. Package Dataset Creation for the \R code used in this calculation.
}
\usage{
hill.shading
}
\description{
Hill shading of the Wood River Valley and surrounding area.
}
\examples{
raster::image(hill.shading, length(hill.shading), col = grey(0:255 / 255), asp = 1,
              axes = FALSE, xlab = "", ylab = "")

}
\keyword{datasets}
