\name{public.parcels}

\alias{public.parcels}

\docType{data}

\title{Public Land Parcels}

\description{
Non-irrigated public land parcels in the Wood River Valley and surrounding areas.
}

\usage{public.parcels}

\format{
An object of \code{SpatialPolygons} class containing 669 \code{Polygons}.
}

\source{
Idaho Department of Water Resources, accessed on November 29, 2014;
derived from Blaine County tax lots and aerial photography
}

\examples{
plot(public.parcels)
print(public.parcels)
}

\keyword{datasets}
