% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{river.reaches}
\alias{river.reaches}
\title{Major River Reaches}
\format{
An object of SpatialLinesDataFrame class containing 22 Lines and a
  data.frame with the following variables:
  \describe{
    \item{Reach}{name of the subreaches measured in U.S. Geological Survey (USGS)
      seepage survey.}
    \item{BigReach}{name of the reaches for which time series targets are available for
      part or all of the calibration period.}
    \item{DrainRiver}{model boundary assignment, either \dQuote{drain} or \dQuote{river}.}
    \item{RchAvg}{estimated average reach gain in cubic meters per day for 1995-2010
      based on a combination of gage data and the USGS seepage survey.}
    \item{BigRAv}{estimated average reach gain in cubic meters per day for 1995-2010
      based on gage data.}
    \item{ReachNo}{reach number identifier.}
    \item{Depth}{estimated average depth in meters of water in reach,
      measured from the air-water interface to the top of the riverbed sediments.}
    \item{BedThk}{estimated thickness in meters of the saturated riverbed sediments.}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
Idaho Department of Water Resources, accessed on June 6, 2015
}
\usage{
river.reaches
}
\description{
Major river reaches of the Wood River Valley, Idaho.
}
\examples{
sp::plot(river.reaches)
str(river.reaches@data)

}
\keyword{datasets}
