% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunWaterBalance.R
\name{RunWaterBalance}
\alias{RunWaterBalance}
\title{Run Water Balance}
\usage{
RunWaterBalance(
  r.grid,
  tr.stress.periods,
  ss.stress.periods = NULL,
  canal.seep = wrv::canal.seep,
  comb.sw.irr = wrv::comb.sw.irr,
  div.gw = wrv::div.gw,
  div.sw = wrv::div.sw,
  div.ww = wrv::div.ww,
  efficiency = wrv::efficiency,
  entity.components = wrv::entity.components,
  et = wrv::et,
  irr.entities = wrv::irr.entities,
  land.surface = wrv::land.surface,
  pod.gw = wrv::pod.gw,
  priority.cuts = wrv::priority.cuts,
  r.canals = wrv::r.canals,
  rs.entities = wrv::rs.entities,
  rs.rech.non.irr = wrv::rs.rech.non.irr,
  verbose = FALSE
)
}
\arguments{
\item{r.grid}{RasterLayer.
Gridded numeric values where NA indicates an \sQuote{inactive} cell in
the top model layer.}

\item{tr.stress.periods}{Date.
Vector of start and end dates for each stress period in the simulation.}

\item{ss.stress.periods}{Date.
Vector of start and end dates for stress periods used to create steady-state conditions.}

\item{canal.seep}{data.frame.
See \code{\link{canal.seep}} dataset for details.}

\item{comb.sw.irr}{data.frame.
See \code{\link{comb.sw.irr}} dataset for details.}

\item{div.gw}{data.frame.
See \code{\link{div.gw}} dataset for details.}

\item{div.sw}{data.frame.
See \code{\link{div.sw}} dataset for details.}

\item{div.ww}{data.frame.
See \code{\link{div.gw}} dataset for details.}

\item{efficiency}{data.frame.
See \code{\link{efficiency}} dataset for details.}

\item{entity.components}{list.
See \code{\link{entity.components}} dataset for details.}

\item{et}{RasterStack.
See \code{\link{et}} dataset for details.}

\item{irr.entities}{SpatialPolygonsDataFrame.
See \code{\link{irr.entities}} dataset for details.}

\item{land.surface}{RasterLayer.
See \code{\link{land.surface}} dataset for details.}

\item{pod.gw}{data.frame.
See \code{\link{pod.gw}} dataset for details.}

\item{priority.cuts}{data.frame.
See \code{\link{priority.cuts}} dataset for details.}

\item{r.canals}{RasterLayer.
See \code{\link{r.canals}} dataset for details.}

\item{rs.entities}{RasterStack.
See \code{\link{rs.entities}} dataset for details.}

\item{rs.rech.non.irr}{RasterStack.
See \code{\link{rs.rech.non.irr}} dataset for details.}

\item{verbose}{logical.
Indicates whether to return summary tables:
  \code{natural.rech}, \code{inciden.rech}, and \code{pumping.rech}.}
}
\value{
Returns an object of class list with the following components:

  (1) Water-table flow data (combines natural and incidental groundwater recharge)
      are stored in \code{areal.rech},
      an object of RasterStack class with raster layers for each model stress period;
      cell values are specified as volumetric flow rates in cubic meters per day.

  (2) Production well pumping data are stored in \code{pod.rech},
      an object of \code{data.frame} class with the following components:
      \describe{
        \item{WMISNumber}{unique number assigned to a water right point of diversion.}
        \item{ss,199501,\dots,201012}{volumetric flow rate, specified for each stress period,
          in cubic meters per day.}
      }

  (3) Natural groundwater recharge data are stored in \code{natural.rech},
      an object of data.frame class with the following components:
      \describe{
        \item{YearMonth}{calendar year and month YYYYMM.}
        \item{Area}{land-surface area of non-irrigated lands, in square meters.}
        \item{ET}{evapotranspiration on non-irrigated lands, in cubic meters per month.}
        \item{Rech}{volumetric flow rate, in cubic meters per month.}
      }

  (4) Incidental groundwater recharge data are stored in \code{inciden.rech},
      an object of data.frame class with the following components:
      \describe{
        \item{EntityName}{name of the irrigation entity.}
        \item{YearMonth}{calendar year and month YYYYMM.}
        \item{SWDiv}{surface-water diversions, in cubic meters per month.}
        \item{SeepFrac}{canal seepage as a fraction of diversions, a dimensionless quantity.}
        \item{CanalSeep}{canal seepage, in cubic meters per month.}
        \item{SWDel}{surface-water delivered to field headgates, in cubic meters per month.}
        \item{area.sw}{area irrigated by only surface water, in square meters.}
        \item{et.sw}{evapotranspiration on lands irrigated by only surface water,
          in cubic meters per month.}
        \item{precip.sw}{precipitation on lands irrigated by only surface water,
          in cubic meters per month.}
        \item{cir.sw}{crop irrigation requirement on lands irrigated by only surface water,
          in cubic meters per month.}
        \item{area.mix}{area irrigated by both surface and groundwater, in square meters.}
        \item{et.mix}{evapotranspiration on lands irrigated by both surface and groundwater,
          in cubic meters per month.}
        \item{precip.mix}{precipitation on lands irrigated by both surface and groundwater,
          in cubic meters per month.}
        \item{cir.mix}{crop irrigation requirement on lands irrigated by both surface and groundwater,
          in cubic meters per month.}
        \item{area.gw}{area irrigated by only groundwater, in square meters.}
        \item{et.gw}{evapotranspiration on lands irrigated by only groundwater,
          in cubic meters per month.}
        \item{precip.gw}{precipitation on lands irrigated by only groundwater,
          in cubic meters per month.}
        \item{cir.gw}{crop irrigation requirement on lands irrigated by only groundwater,
          in cubic meters per month.}
        \item{Eff}{irrigation efficiency, a dimensionless quantity.}
        \item{GWDiv}{recorded groundwater diversions, in cubic meters per month.}
        \item{WWDiv}{inflow to municipal wastewater treatment plants, in cubic meters per month.}
        \item{hg.sw}{surface-water delivered to field headgates on lands irrigated by only surface water,
          in cubic meters per month.}
        \item{hg.mix}{surface-water delivered to field headgates on lands irrigation by both surface and groundwater,
          in cubic meters per month.}
        \item{rech.sw}{incidental groundwater recharge beneath lands irrigated by only surface water,
          in cubic meters per month.}
        \item{gw.dem.mix}{groundwater demand on lands irrigated by both surface and groundwater,
          in cubic meters per month.}
        \item{gw.div.est}{calculated groundwater diversions, in cubic meters per month.}
        \item{rech.mix}{incidental groundwater recharge beneath lands irrigated by both surface and groundwater,
          in cubic meters per month.}
        \item{gw.only}{groundwater demand on lands irrigated by only groundwater in entities with
          lands also irrigated by both surface and groundwater, in cubic meters per month.}
        \item{rech.muni}{incidental groundwater recharge beneath entities with
          lands irrigated by only groundwater and lands irrigated by both surface and groundwater,
          in cubic meters per month.}
        \item{gw.dem.gw}{groundwater demand on lands irrigated by only groundwater in
          entities without surface-water irrigation, in cubic meters per month.}
        \item{rech.gw}{incidental groundwater recharge beneath lands irrigated by only groundwater,
          in cubic meters per month.}
        \item{area.model}{area of the irrigation entity that is located in the model domain,
          in square meters.}
      }
     Volumetric flow rates are calculated for their respective area in
     the irrigation entity---not just that part overlying the model area.
     Flow rate values are given this way in order to facilitate with quality assurance of
     the water-budget calculation.
     To calculate a simulated volumetric-flow rate: divide the flow rate by the affected area,
     and then multiply this value by the area of the irrigation entity that is located in
     the model domain (\code{area.model}).

  (5) Well pumping data are also stored in \code{pumping.rech} (see \code{pod.rech} component),
      an object of data.frame class with the following components:
      \describe{
        \item{WMISNumber}{unique number assigned to a water right point of diversion.}
        \item{YearMonth}{calendar year and month YYYYMM.}
        \item{Pumping}{volumetric rate of pumping, in cubic meters per month.}
      }
}
\description{
This function estimates natural and incidental groundwater recharge at the water table,
and pumping demand at production wells.
A water-balance approach is used to calculate these volumetric flow rate estimates,
where positive values are flow into the aqufer system (groundwater recharge),
and negative values are flow out of the system (groundwater discharge).
}
\examples{
\dontrun{# see Appendix A. Package Introduction}

}
\seealso{
\code{\link{UpdateWaterBudget}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center

  J. Sukow and M. McVay, Idaho Department of Water Resources
}
\keyword{manip}
