\name{sensitivity}

\alias{sensitivity}

\docType{data}

\title{PEST Sensitivity}

\description{
Calibrated parameter values and composite sensitivities generated by PEST.
}

\usage{sensitivity}

\format{
A \code{data.frame} object with 336 records and the following variables:
\describe{
  \item{parameter.desc}{is a description of the parameter.}
  \item{ID}{is a unique identifier within the parameter type, such as an identifier for a pilot point or irrigation entity.}
  \item{units}{is the parameter units.}
  \item{start.value}{is the starting parameter value prior to model calibration.}
  \item{lower.bound}{is the lower bound placed on the parameter value during the model-calibration process.}
  \item{upper.bound}{is the upper bound placed on the parameter value during the model-calibration process.}
  \item{parameter.name}{is the \href{http://www.pesthomepage.org/}{PEST} parameter name.}
  \item{group}{is the PEST parameter group.}
  \item{value}{is the calibrated parameter value estimated by PEST.}
  \item{comp.sens}{is the composite sensitivity generated during the final iteration of PEST.}
  \item{rel.comp.sens}{is the relative composite sensitivity.}
}
}

\source{
Idaho Department of Water Resources, accessed on January 15, 2016
}

\seealso{
\code{\link{pilot.points}}, \code{\link{irr.entities}}, \code{\link{river.reaches}}, \code{\link{drains}}, \code{\link{tributaries}}
}

\examples{
str(sensitivity)
}

\keyword{datasets}
