% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soils}
\alias{soils}
\title{Soil Units}
\format{
An object of SpatialPolygonsDataFrame class containing 718 Polygons and a
  data.frame with the following variables:
  \describe{
    \item{GroupSymbol}{soil class identifier}
    \item{SoilLayer}{identifier used to differentiate the soil data source
      used to create the soils map.
      Data sources are either \sQuote{USCS} or \sQuote{STATSGO},
      the NRCS State Soil Geographic Data Base.}
    \item{SoilClass}{description of the soil class.}
    \item{MinRate}{lower percolation rate limit for the soil class, in meters per month.}
    \item{MaxRate}{upper percolation rate limit for the soil class, in meters per month.}
    \item{PercolationRate}{percolation rate in meters per month.}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
Idaho Department of Water Resources, accessed on April 22, 2015
}
\usage{
soils
}
\description{
Representation of mapped surficial soil units created by the Idaho Office of the
National Resource Conservation Service (NRCS).
Soils have been assigned a percolation rate based on the average,
saturated hydraulic conductivity of the soils as classified using the
Unified Soil Classification System (USCS).
}
\examples{
sp::spplot(soils, "PercolationRate")
str(soils@data)

}
\keyword{datasets}
