\name{wetlands}

\alias{wetlands}

\docType{data}

\title{Wetlands}

\description{
Wetlands in the Wood River Valley and surrounding areas.
}

\usage{wetlands}

\format{
An object of \code{SpatialPolygons} class containing 3,024 \code{Polygons}.
}

\source{
U.S. Fish and Wildlife Service National Wetlands Inventory, accessed on April 2, 2014
}

\examples{
plot(wetlands, col = "#CCFFFF", border = "#3399CC", lwd = 0.5)
print(wetlands)
}

\keyword{datasets}
