% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zone.properties}
\alias{zone.properties}
\title{Hydraulic Properties of Hydrogeologic Zones}
\format{
An object of \code{data.frame} class with the following variables:
  \describe{
    \item{ID}{numeric identifier for the hydrogeologic zone.}
    \item{name}{name of the hydrogeologic zone.}
    \item{vani}{vertical anisotropy, a dimensionless quantity.}
    \item{sc}{storage coefficient, a dimensionless quantity.}
    \item{sy}{specific yield, a dimensionless quantity.}
    \item{hk}{horizontal hydraulic conductivity in meters per day.}
    \item{ss}{specific storage in inverse meter.}
  }
}
\usage{
zone.properties
}
\description{
Estimates of the hydraulic properties for each hydrogeologic zone.
}
\examples{
str(zone.properties)

}
\references{
Bartolino, J.R., and Adkins, C.B., 2012, Hydrogeologic framework of the
  Wood River Valley aquifer system, south-central Idaho:
  U.S. Geological Survey Scientific Investigations Report 2012-5053, 46 p.,
  available at \url{https://pubs.usgs.gov/sir/2012/5053/}.
}
\keyword{datasets}
