% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alluvium.extent}
\alias{alluvium.extent}
\title{Extent of Alluvium Unit}
\format{An object of SpatialPolygonsDataFrame class containing 1 Polygons.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Extent defined by Bartollino and Adkins (2012, Plate 1).
}
\usage{
alluvium.extent
}
\description{
Estimated extent of alluvium unit in the Wood River Valley aquifer system.
}
\examples{
sp::plot(alluvium.extent, col = "#BFA76F")
str(alluvium.extent)

}
\references{
Bartolino, J.R., and Adkins, C.B., 2012,
  Hydrogeologic framework of the Wood River Valley aquifer system, south-central Idaho:
  U.S. Geological Survey Scientific Investigations Report 2012-5053, 46 p.,
  available at \url{https://pubs.usgs.gov/sir/2012/5053/}.
}
\keyword{datasets}
