% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bellevue.wwtp.ponds}
\alias{bellevue.wwtp.ponds}
\title{Bellevue Waste Water Treatment Plant Ponds}
\format{
An object of SpatialPolygonsDataFrame class containing 1 Polygons.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
Idaho Department of Water Resources, accessed on December 11, 2014
}
\usage{
bellevue.wwtp.ponds
}
\description{
Location of the Bellevue Waste Water Treatment Plant ponds.
}
\examples{
sp::plot(bellevue.wwtp.ponds)

}
\keyword{datasets}
