% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass.canal}
\alias{bypass.canal}
\title{Bypass Canal}
\format{An object of SpatialLines class containing 4 Lines.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Idaho Department of Water Resources, accessed on January 15, 2015
}
\usage{
bypass.canal
}
\description{
Location of the Bypass Canal in the Wood River Valley.
}
\examples{
sp::plot(bypass.canal)

}
\keyword{datasets}
