% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{canals}
\alias{canals}
\title{Canal Systems}
\format{An object of SpatialLinesDataFrame class containing
  113 Lines and a data.frame with the following variable:
  \describe{
    \item{EntityName}{name of the irrigation entity served by the canal system.}
    \item{Name}{local canal name}
  }}
\source{
Idaho Department of Water Resources, accessed on November 29, 2014
}
\usage{
canals
}
\description{
Canal systems in the Wood River Valley and surrounding areas.
}
\examples{
sp::plot(canals, col = "#3399CC")
str(canals@data)

}
\seealso{
\code{\link{r.canals}}, \code{\link{canal.seep}}
}
\keyword{datasets}

