% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{div.ret.exch}
\alias{div.ret.exch}
\title{Diversions, Returns, and Exchange Wells}
\format{An object of SpatialPointsDataFrame class containing 117 points with the
  following variables:
  \describe{
    \item{Name}{local name for the diversion/return site.}
    \item{Type}{data type, either \dQuote{Diversion}, \dQuote{Return}, or
      \dQuote{Exchange well inflow}.}
    \item{LocSource}{data source}
    \item{Big}{corresponding river reach}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Idaho Department of Water Resources, accessed on June 5, 2015
}
\usage{
div.ret.exch
}
\description{
Location of streamflow diversions, irrigation canal or pond returns, and
exchange well returns.
}
\examples{
sp::plot(div.ret.exch)
str(div.ret.exch@data)

}
\keyword{datasets}
