% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{div.sw}
\alias{div.sw}
\title{Surface-Water Diversions}
\format{An object of class data.frame with 15,550 records and the following variables:
  \describe{
    \item{YearMonth}{year and month during which diversions were recorded,
      with a required date format of YYYYMM.}
    \item{Diversion}{name of the surface-water diversion.}
    \item{Reach}{river subreach from which the water is diverted.}
    \item{BigReach}{river reach from which the water is diverted.}
    \item{EntityName}{name of the irrigation entity which the diversion supplies water.}
    \item{SWDiv}{volume of water diverted during the month, in cubic meters.}
  }}
\source{
Idaho Department of Water Resources, accessed on December 11, 2014;
  compiled data records from Water District 37 and 37M, City of Hailey,
  City of Bellevue, City of Ketchum, and Sun Valley Water and Sewer District.
}
\usage{
div.sw
}
\description{
Surface-water diversions recorded by Water District 37 or municipal water providers.
}
\examples{
str(div.sw)

}
\keyword{datasets}
