% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drains}
\alias{drains}
\title{Drain Boundaries at Stanton Crossing and Silver Creek}
\format{
An object of SpatialPolygonsDataFrame class containing a set of 2 Polygons and
  a data.frame with the following variable:
  \describe{
    \item{Name}{identifier for the polygon.}
    \item{cond}{drain conductance in square meters per day.}
    \item{elev}{drain threshold elevation in meters above the
      North American Vertical Datum of 1988 (NAVD 88).}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
U.S. Geological Survey, accessed on March 27, 2015;
  a Keyhole Markup Language (\href{https://en.wikipedia.org/wiki/Kml}{KML}) file created in
  \href{https://www.google.com/earth/}{Google Earth} with polygons drawn by hand in
  areas of known drains.
}
\usage{
drains
}
\description{
Polygons used to define the locations of drain boundaries in the model domain.
The polygons clip the line segments along the aquifer boundary (see \code{\link{alluvium.extent}}),
and model cells intersecting these clipped-line segments are defined as boundary cells.
}
\examples{
str(drains)

sp::plot(drains, border = "red")
sp::plot(alluvium.extent, add = TRUE)

}
\keyword{datasets}
