% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{et}
\alias{et}
\title{Evapotranspiration}
\format{An object of RasterStack class containing 192 RasterLayer objects,
  one layer for each month in the 1995-2010 time period.
  Each cell on a layers surface grid represents the monthly depth of ET in meters.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Idaho Department of Water Resources, accessed on November 17, 2014
}
\usage{
et
}
\description{
Evapotranspiration (ET) in the Wood River Valley and surrounding areas.
Defined as the amount of water lost to the atmosphere via direct evaporation,
transpiration by vegetation, or sublimation from snow covered areas.
}
\examples{
sp::plot(et[["199505"]])
print(et)

}
\seealso{
\code{\link{et.method}}
}
\keyword{datasets}
