% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSeasonalMult.R
\name{GetSeasonalMult}
\alias{GetSeasonalMult}
\title{Get Seasonal Multiplier}
\usage{
GetSeasonalMult(x, reduction, d.in.mv.ave, fixed.dates)
}
\arguments{
\item{x}{data.frame.
Time series data (observations) with components of class Date and numeric.}

\item{reduction}{numeric.
Signal amplitude reduction, a dimensionless quantity.
Its magnitude should be greater than or equal to 1;
where a value of 1 indicates no reduction in the signal amplitude.}

\item{d.in.mv.ave}{numeric.
Number of days in the moving average subset.}

\item{fixed.dates}{Date.
Vector of equally spaced dates, these are the fixed locations where the moving average is calculated.
The final date is neglected.}
}
\value{
Returns an object of class data.frame with the following variables:
  \describe{
    \item{\code{names(x)[1]}}{start date for each season in \code{fixed.dates}.}
    \item{multiplier}{seasonal scaling index}
  }
}
\description{
This function determines the seasonal fraction of the long-term mean value.
}
\details{
A simple moving average is first calculated at dates specified in \code{fixed.dates}
  using past observational data in \code{x}
 (such as the previous 9-months of stage data recorded at a streamgage).
  The seasonal average of the moving average is then passed through a signal amplitude reduction algorithm.
  The reduced values are then divided by the mean of the seasonal reduced data to give
  the seasonal fraction of the mean (also known as the seasonal scaling index).
}
\examples{
obs <- dataRetrieval::readNWISdata(sites = "13139510", parameterCd = "00060",
                                   startDate = "1992-01-01", endDate = "2011-01-01")
obs <- obs[, c("dateTime", "X_00060_00003")]
obs[, 1] <- as.Date(obs[, 1])

fixed.dates <- seq(as.Date("1995-01-01"), as.Date("2011-01-01"), "1 month")
d <- GetSeasonalMult(obs, 2, 273.932, fixed.dates)
str(d)

}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center

  A.H. Wylie and J. Sukow, Idaho Department of Water Resources
}
\keyword{manip}
