% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{idaho}
\alias{idaho}
\title{U.S. State of Idaho}
\format{An object of SpatialPolygons class containing 1 Polygons.
  Cartographic boundary at 5m (1:5,000,000) resolution.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
U.S. Department of Commerce, U.S. Census Bureau,
  Geography Division/Cartographic Products Branch.
  A simplified representation of the State of Idaho from the 2014 Census Bureau's
  MAF/\href{http://www.census.gov/geo/maps-data/data/tiger.html}{TIGER} geographic database.
}
\usage{
idaho
}
\description{
Boundary of Idaho, a state in the northwestern region of the United States.
}
\examples{
sp::plot(idaho, col = "#EAE2CF", border = "#BFA76F", lwd = 0.5)
print(idaho)

}
\keyword{datasets}

