% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irr.entities}
\alias{irr.entities}
\title{Irrigation Entities}
\format{
An object of SpatialPolygonsDataFrame class containing 235 Polygons and
  a data.frame with the following variables:
  \describe{
    \item{EntityName}{name of the irrigation entity served by a group of diversions.}
    \item{Source}{water source, either \dQuote{Mixed} for a mixture of surface water and groundwater,
      \dQuote{SW Only} for surface-water only, or \dQuote{GW Only} for groundwater only.}
    \item{EntitySrce}{concatenation of the \code{EntityName} and \code{Source} character strings.}
    \item{PrecipZone}{name of the precipitation zone.
      See \code{\link{precip.zones}} dataset for details.}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
Idaho Department of Water Resources (IDWR), accessed on December 11, 2014;
  derived from IDWR water rights database, Blaine County tax lot data,
  and IDWR irrigated land classification files.
}
\usage{
irr.entities
}
\description{
Delineation of areas served by a group of surface-water and (or) groundwater diversions.
}
\examples{
sp::plot(irr.entities)
print(irr.entities)

}
\keyword{datasets}
