% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irr.lands}
\alias{irr.lands}
\title{Irrigated Lands}
\format{An object of list class with 7 SpatialPolygonsDataFrame components.
  The data.frame associated with each SpatialPolygons object has the following variable:
  \describe{
    \item{Status}{status of land during the year reviewed,
      either \dQuote{irrigated}, \dQuote{semi-irrigated}, or \dQuote{non-irrigated}.}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Idaho Department of Water Resources, accessed on November 17, 2014;
  polygons derived from U.S. Department of Agriculture Common Land Unit polygons
  with some refinement of polygons.
  Irrigation status interpreted using satellite imagery and aerial photography.
}
\usage{
irr.lands
}
\description{
Classification of irrigated lands in the Wood River Valley and surrounding areas;
available for years 1996, 2000, 2002, 2006, 2008, 2009, and 2010.
}
\examples{
sp::spplot(irr.lands[["2010"]], "Status")
print(irr.lands)

}
\seealso{
\code{\link{irr.lands.year}}
}
\keyword{datasets}
