% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{irr.lands.year}
\alias{irr.lands.year}
\title{Irrigation Lands for a Given Year}
\format{
An object of data.frame class with 16 records and the following variables:
  \describe{
    \item{Year}{year with a required date format of YYYY.}
    \item{IL_Year}{substitute year with a required date format of YYYY.}
  }
}
\source{
Idaho Department of Water Resources, accessed on April 25, 2014
}
\usage{
irr.lands.year
}
\description{
Annual land classification for irrigation practices is only available for select years.
For missing years, this dataset provides substitute years when land-classification was
available (see \code{\link{irr.lands}}).
}
\examples{
str(irr.lands.year)

}
\keyword{datasets}
