% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kriging.zones}
\alias{kriging.zones}
\title{Kriging Zones}
\format{An object of SpatialPolygonsDataFrame class containing 18 Polygons and a
data.frame with the following variables:
  \describe{
    \item{Zone}{kriging zone, interpolation in this zone is based on the
      parameter values of pilot points located within this zone.}
    \item{Layer}{model layer}
    \item{Name}{local name}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Idaho Department of Water Resources, accessed on June 11, 2015
}
\usage{
kriging.zones
}
\description{
Location of kriging zones in the Wood River Valley, used in parameter estimation.
}
\examples{
sp::plot(kriging.zones)
str(kriging.zones@data)

}
\seealso{
\code{\link{pilot.points}}
}
\keyword{datasets}
