% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lakes}
\alias{lakes}
\title{Lakes and Reservoirs}
\format{An object of SpatialPolygonsDataFrame class containing 55 Polygons.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Idaho Department of Water Resources
  (\href{https://research.idwr.idaho.gov/index.html#GIS-Data}{IDWR}),
  accessed on April 2, 2014
}
\usage{
lakes
}
\description{
Lakes and reservoirs of the Wood River Valley and surrounding areas.
}
\examples{
sp::plot(lakes, col = "#CCFFFF", border = "#3399CC", lwd = 0.5)
str(lakes@data)

}
\keyword{datasets}
