% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{major.roads}
\alias{major.roads}
\title{Major Roads}
\format{
An object of SpatialLinesDataFrame class containing 475 Lines.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
Idaho Department of Water Resources
  (\href{https://research.idwr.idaho.gov/index.html#GIS-Data}{IDWR}),
  accessed on October 20, 2015
}
\usage{
major.roads
}
\description{
Major roads in the Wood River Valley and surrounding areas.
}
\examples{
sp::plot(major.roads)
str(major.roads@data)

}
\keyword{datasets}
