% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{map.labels}
\alias{map.labels}
\title{Map Labels}
\format{
An object of SpatialPointsDataFrame class containing
  40 points with the following variables:
  \describe{
    \item{label}{text to be written.}
    \item{cex}{character expansion factor}
    \item{col,font}{color and font to be used, respectively.}
    \item{srt}{string rotation in degrees.}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
Best estimates of map label locations.
}
\usage{
map.labels
}
\description{
Map labels in the Wood River Valley and surrounding areas.
}
\examples{
sp::plot(map.labels, col = "red")
lab <- cbind(map.labels@coords, map.labels@data)
for (i in seq_len(nrow(lab))) {
  text(lab$x[i], lab$y[i], labels = lab$label[i], cex = lab$cex[i],
       col = lab$col[i], font = lab$font[i], srt = lab$srt[i])
}

}
\keyword{datasets}
