% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{misc.seepage}
\alias{misc.seepage}
\title{Direct Recharge from Miscellaneous Seepage Sites}
\format{An object of data.frame class with 2 records and the following variables:
  \describe{
    \item{RechSite}{name of the recharge site, see \code{\link{bellevue.wwtp.ponds}} and
      \code{\link{bypass.canal}} datasets.}
    \item{199501,\dots,201012}{monthly volume of recharge during a stress period, in cubic meters.
      The variable name is specified as year and month.}
  }}
\source{
Idaho Department of Water Resources, accessed on January 15, 2015
}
\usage{
misc.seepage
}
\description{
Direct recharge from miscellaneous seepage sites in the Wood River Valley, Idaho.
}
\examples{
str(misc.seepage)

}
\keyword{datasets}
