% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs.wells.head}
\alias{obs.wells.head}
\title{Hydraulic Heads in Observation Wells}
\format{An object of class data.frame with 3,477 records and the following variables:
  \describe{
    \item{PESTNAME}{unique well identifier for PEST.}
    \item{DateTime}{date and time when the measurement was recorded.}
    \item{Head}{groundwater-level measurement (hydraulic head)
      in meters above the North American Vertical Datum of 1988 (NAVD 88).}
  }}
\source{
Idaho Department of Water Resources, accessed on June 26, 2015
}
\usage{
obs.wells.head
}
\description{
Hydraulic-head (groundwater-level) measurements recorded in observation wells in the
Wood River Valley aquifer system.
Values are used as observations during the parameter estimation process.
}
\examples{
str(obs.wells.head)

}
\seealso{
\code{\link{obs.wells}}
}
\keyword{datasets}
