% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obs.wells}
\alias{obs.wells}
\title{Observation Wells}
\format{An object of SpatialPointsDataFrame class containing 776 points
  with the following variables:
  \describe{
    \item{id}{unique well identifier used in this study.}
    \item{SiteNo}{unique well identifier within the
      National Water Information System (NWIS).}
    \item{SITEIDIDWR}{unique well identifier within the
      Idaho Department of Water Resources (IDWR) hydrologic database.}
    \item{WELLNUMBER}{USGS or IDWR site name for the well.}
    \item{PESTNAME}{unique well identifier for PEST.}
    \item{METHODDRIL}{drilling method}
    \item{TOTALDEPTH}{depth at which drilling stopped, in feet.}
    \item{OPENINGMIN}{top of the screened interval, in feet.}
    \item{OPENINGMAX}{bottom of the screened interval, in feet.}
    \item{COMPLETION}{date on which the well drilling and construction stopped.}
    \item{WCWELLID}{well construction well identifier.}
    \item{ALTITUDE}{land surface elevation, in feet.}
    \item{ALTMETHOD}{method for obtaining the land surface elevation.}
    \item{XYMETHOD}{method of obtaining the spatial coordinates.}
    \item{BASINNO}{basin number}
    \item{COUNTYNAME}{Idaho county name}
    \item{TWPRGE}{township and range the well is located in.}
    \item{SITENAME}{local name for well.}
    \item{desc}{description of well type.}
    \item{TopOpen1}{depth to the top of the first open interval in a groundwater well,
      in meters below land surface.}
    \item{BotOpen1}{depth to the bottom of the first open interval in a groundwater well,
      in meters below land surface.}
    \item{TopOpen2}{not applicable}
    \item{BotOpen2}{not applicable}
  }}
\source{
Idaho Department of Water Resources well construction database,
  accessed on June 29, 2015
}
\usage{
obs.wells
}
\description{
Observation wells in the Wood River Valley aquifer system.
}
\examples{
sp::plot(obs.wells)
str(obs.wells@data)

}
\seealso{
\code{\link{obs.wells.head}}
}
\keyword{datasets}

