% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pilot.points}
\alias{pilot.points}
\title{Pilot Points}
\format{
An object of SpatialPointsDataFrame class containing 106 points
  with the following variables:
  \describe{
    \item{Layer}{model layer}
    \item{Zone}{kriging-zone identifier}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
Idaho Department of Water Resources, accessed on June 11, 2015
}
\usage{
pilot.points
}
\description{
Location of pilot points in the model domain.
}
\examples{
sp::plot(pilot.points)
str(pilot.points@data)

}
\seealso{
\code{\link{kriging.zones}}
}
\keyword{datasets}
