% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pod.gw}
\alias{pod.gw}
\title{Points of Diversion for Groundwater}
\format{An object of class data.frame with 1,755 records and the following variables:
  \describe{
    \item{WMISNumber}{unique number assigned to a water right point of diversion.}
    \item{WaterRight}{number identifying a specific authorization to use
      water in a prescribed manner.}
    \item{EntityName}{name of the irrigation entity the point of diversion is assigned to.}
    \item{EntitySrce}{source of water for an irrigation entity.
      Possible sources of water include surface water, groundwater and mixed.
      Mixed source entities derive water from both groundwater and surface water.}
    \item{Pdate}{priority date, the date the water right was established.}
    \item{IrrRate}{irrigation rate in cubic meters per day,
      the maximum permitted water use rate associated with a water right.}
  }}
\source{
Idaho Department of Water Resources water rights database,
  accessed on December 11, 2014
}
\usage{
pod.gw
}
\description{
Points of diversion for groundwater within the Wood River Valley model study area.
}
\examples{
summary(pod.gw)

}
\seealso{
\code{\link{pod.wells}}
}
\keyword{datasets}
