% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{public.parcels}
\alias{public.parcels}
\title{Public Land Parcels}
\format{An object of SpatialPolygons class containing 669 Polygons.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Idaho Department of Water Resources, accessed on November 29, 2014;
  derived from Blaine County tax lots and aerial photography.
}
\usage{
public.parcels
}
\description{
Non-irrigated public land parcels in the Wood River Valley and surrounding areas.
}
\examples{
sp::plot(public.parcels)
print(public.parcels)

}
\keyword{datasets}
