% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r.canals}
\alias{r.canals}
\title{Rasterized Canals}
\format{An object of RasterLayer class with indexed cell values linked to
  a raster attribute table (RAT).
  The RAT is a \code{data.frame} with the following components:
  \describe{
    \item{ID}{cell index}
    \item{COUNT}{frequency of the cell index in the raster grid.}
    \item{EntityName}{name of the irrigation entity served by the canal system.}
  }}
\source{
Calculated by transferring the \code{\link{canals}} dataset to grid cells in the
  \code{\link{land.surface}} dataset using the \code{rasterize} function;
  see the appendix C. Package Dataset Creation for the \R code used in this calculation.
}
\usage{
r.canals
}
\description{
Canal systems of the Wood River Valley and surrounding areas transferred to raster cells.
}
\examples{
sp::plot(r.canals)
print(levels(r.canals)[[1]])

}
\keyword{datasets}

