% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rs.entities}
\alias{rs.entities}
\title{Rasterized Monthly Irrigation Entities}
\format{
An object of RasterStack class containing a 192 RasterLayer objects,
  one layer for each month in the 1995-2010 time period.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
  For each raster layer, indexed cell values are linked to a raster attribute table (RAT).
  The RAT is an object of data.frame class with the following components:
  \describe{
    \item{ID}{cell index}
    \item{COUNT}{frequency of the cell index in the raster grid.}
    \item{EntityName}{name of the irrigation entity served by a group of diversions.}
  }
}
\source{
Calculated by transferring the \code{\link{entity.components}} dataset to grid cells
  in the \code{\link{land.surface}} dataset using the rasterize function;
  see the appendix C. Package Dataset Creation for the \R code used in this calculation.
}
\usage{
rs.entities
}
\description{
Irrigation entities of the Wood River Valley and surrounding areas transferred to raster cells.
}
\examples{
names(rs.entities)
sp::plot(rs.entities[["199507"]])
print(levels(rs.entities[["199507"]])[[1]])

}
\keyword{datasets}
