% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sensitivity}
\alias{sensitivity}
\title{PEST Sensitivity}
\format{
An object of data.frame class with 336 records and the following variables:
  \describe{
    \item{parameter.desc}{parameter description}
    \item{ID}{unique identifier within the parameter type,
      such as an identifier for a pilot point or irrigation entity.}
    \item{units}{parameter units}
    \item{start.value}{starting parameter value prior to model calibration.}
    \item{lower.bound}{lower bound placed on the parameter value
      during the model-calibration process.}
    \item{upper.bound}{upper bound placed on the parameter value
      during the model-calibration process.}
    \item{parameter.name}{\href{http://www.pesthomepage.org/}{PEST} parameter name}
    \item{group}{PEST parameter group}
    \item{value}{calibrated parameter value estimated by PEST.}
    \item{comp.sens}{composite sensitivity generated during the final iteration of PEST.}
    \item{rel.comp.sens}{relative composite sensitivity}
  }
}
\source{
Idaho Department of Water Resources, accessed on January 15, 2016
}
\usage{
sensitivity
}
\description{
Calibrated parameter values and composite sensitivities generated by PEST.
}
\examples{
str(sensitivity)

}
\seealso{
\code{\link{pilot.points}}, \code{\link{irr.entities}},
  \code{\link{river.reaches}}, \code{\link{drains}}, \code{\link{tributaries}}
}
\keyword{datasets}
