% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streamgages}
\alias{streamgages}
\title{Streamgages}
\format{
An object of SpatialPointsDataFrame class containing 9 points and a
  data.frame with the following variable:
  \describe{
    \item{SiteNo}{unique site number for the streamgage.}
    \item{SiteName}{official name of the streamgage.}
  }
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
National Water Information System (\href{https://waterdata.usgs.gov/nwis}{NWIS}),
  accessed on May 29, 2015.
}
\usage{
streamgages
}
\description{
Select streamgages in the Wood River Valley.
}
\examples{
str(streamgages)

}
\keyword{datasets}
