% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streams.rivers}
\alias{streams.rivers}
\title{Streams and Rivers}
\format{An object of SpatialLinesDataFrame class containing 581 Lines.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).}
\source{
Idaho Department of Water Resources
  (\href{https://research.idwr.idaho.gov/index.html#GIS-Data}{IDWR}),
  accessed on April 2, 2014
}
\usage{
streams.rivers
}
\description{
Streams and rivers of the Wood River Valley and surrounding areas.
}
\examples{
sp::plot(streams.rivers, col = "#3399CC")
str(streams.rivers@data)

}
\keyword{datasets}
