% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subreach.recharge}
\alias{subreach.recharge}
\title{Stream-Aquifer Flow Exchange Along River Subreaches}
\format{An object of data.frame class with 19 records and the following variables:
  \describe{
    \item{ReachNo}{subreach number identifier}
    \item{Reach}{name of the subreach.}
    \item{BigReachNo}{reach number identifier}
    \item{BigReach}{name of the reach.}
    \item{Aug}{estimated volumetric flow rate measured during August 2012,
      in cubic meters per day.}
    \item{Oct}{estimated volumetric flow rate measured during October 2012,
      in cubic meters per day.}
    \item{Mar}{estimated volumetric flow rate measured during March 2013,
      in cubic meters per day.}
  }}
\source{
Flow values calculated from \code{\link{seepage.study}} data.
}
\usage{
subreach.recharge
}
\description{
Stream-aquifer flow exchange along river subreaches specified as aquifer recharge.
Values used as observations in parameter estimation.
}
\details{
A positive stream-aquifer flow exchange indicates aquifer recharge
  (also know as a losing river subreach).
}
\examples{
str(subreach.recharge)

}
\keyword{datasets}
