% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swe}
\alias{swe}
\title{Snow Water Equivalent}
\format{
An object of data.frame class with 366 records and the following variables:
  \describe{
    \item{MonthDay}{month and day, with a required date format of \code{MMDD}.}
    \item{Choco}{daily SWE recorded at the Chocolate Gulch snow telemetry (SNOTEL)
      weather station.}
    \item{Hailey}{daily SWE recorded at the Hailey Ranger Station at Hailey
      hydrometeorological automated data system (HADS) weather station.}
    \item{Picabo}{daily SWE recorded at the Picabo PICI HADS weather station.}
  }
}
\source{
Idaho Department of Water Resources, accessed on April 24, 2015
}
\usage{
swe
}
\description{
Average daily snow water equivalent (SWE) at weather stations in the
Wood River Valley and surrounding areas.
}
\examples{
str(swe)

}
\seealso{
\code{\link{weather.stations}}, \code{\link{precip.zones}},
  \code{\link{precipitation}}
}
\keyword{datasets}
