% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tributary.streams}
\alias{tributary.streams}
\title{Tributary Streams}
\format{
An object of SpatialLinesDataFrame class containing 88 Lines.
  Geographic coordinates are in units of meters, in conformance with the
  North American Datum of 1983 (NAD 83), and placed in the
  Idaho Transverse Mercator projection (\href{https://www.idwr.idaho.gov/GIS/IDTM/}{IDTM}).
}
\source{
Idaho Department of Water Resources, accessed on June 1, 2015
}
\usage{
tributary.streams
}
\description{
Tributary streams of the upper Wood River Valley and surrounding areas.
}
\examples{
sp::plot(tributary.streams, col = "#3399CC")
str(tributary.streams@data)

}
\keyword{datasets}
