% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateWaterBudget.R
\name{UpdateWaterBudget}
\alias{UpdateWaterBudget}
\title{Update Water Budget}
\usage{
UpdateWaterBudget(dir.run, id, qa.tables = c("none", "si", "english"),
  ss.interval = NULL, iwelcb = ifelse(interactive(), 50L, 0L),
  canal.seep = wrv::canal.seep, efficiency = wrv::efficiency,
  gage.disch = wrv::gage.disch, pod.wells = wrv::pod.wells,
  tributaries = wrv::tributaries, ...)
}
\arguments{
\item{dir.run}{character.
Path name of the directory to read/write model files.}

\item{id}{character.
Short identifier (file name) for model files.}

\item{qa.tables}{character.
Indicates if quality assurance tables are written to disk;
by default "none" of these tables are written.
Values of "si" and "english" indicate that table values are written in
metric and English units, respectively.}

\item{ss.interval}{Date or character.
Vector of length 2 specifying the start and end dates for the period used to
represent steady-state boundary conditions.
That is, recharge values for stress periods coinciding with this time period are
averaged and used as steady-state boundary conditions.
The required date format is YYYY-MM-DD.
This argument overrides the \code{ss.stress.periods} object in the \file{model.rda} file,
see \sQuote{Details} section for additional information.}

\item{iwelcb}{integer.
A flag and unit number.
If equal to zero, the default, cell-by-cell flow terms resulting from
conditions in the MODFLOW Well Package will not be written to disk.
A value of 0 is appropriate for model calibration,
where MODFLOW run times are kept as short as possible.
If greater than zero, the cell-by-cell flow terms are written to disk.
See the MODFLOW Name File (\file{*.nam}) for the unit number associated with
the budget file (\file{*.bud}).
The default value is 50 (value specified in the \code{\link{WriteModflowInput}} function)
if \R is being used interactively and 0 otherwise.}

\item{canal.seep}{data.frame.
See \code{\link{canal.seep}} dataset for details.}

\item{efficiency}{data.frame.
See \code{\link{efficiency}} dataset for details.}

\item{gage.disch}{data.frame.
See \code{\link{gage.disch}} dataset for details.}

\item{pod.wells}{SpatialPointsDataFrame.
See \code{\link{pod.wells}} dataset for details.}

\item{tributaries}{SpatialPolygonsDataFrame.
See \code{\link{tributaries}} dataset for details.}

\item{...}{Arguments to be passed to \code{\link{RunWaterBalance}}, such as evapotranspiration \code{et}.}
}
\value{
Returns an object of difftime class, the runtime for this function.
  Used for the side-effect of files written to disk.

  A MODFLOW Well Package file \file{<id>.wel} is always written to disk; whereas,
  parameter estimation files \file{seep.csv}, \file{eff.csv}, and \file{trib.csv}, and
  a script file \file{UpdateBudget.bat}, are only written if they do not already exist.
  The script file may be used to automate the execution of this function from a
  file manager (such as, Windows Explorer).

  The \file{seep.csv} file stores as tabular data the canal seepage fraction for
  each of the irrigation entities.
  Its character and numeric data fields are delimited by commas (a comma-separated-value [CSV] file).
  The first line is reserved for field names \dQuote{EntityName} and \dQuote{SeepFrac}.

  The \file{eff.csv} file stores as tabular data the irrigation efficiency for
  each of the irrigation entities.
  Its character and numeric data fields are delimited by commas.
  The first line is reserved for field names \dQuote{EntityName} and \dQuote{Eff}.

  The \file{trib.csv} file stores as tabular data the underflow boundary conditions for
  each tributary basin.
  Its character and numeric data fields are delimited by commas.
  The first line is reserved for field names \dQuote{Name} and \dQuote{Value}.
  Data records include a long-term mean flow multiplier for
  each of the tributary basins (name is the unique identifier for the tributary),
  a record for the amplitude reduction (\code{reduction}), and
  a record for the number of days in the moving average (\code{d.in.mv.ave}).

  If the \code{qa.tables} argument is specified as either \dQuote{si} or \dQuote{english},
  quality assurance tables are written to disk as CSV files (\file{qa-*.csv}).
  Volumetric flow rate data within these tables is described in the
  \sQuote{Value} section of the \code{\link{RunWaterBalance}} function;
  see returned list components \code{natural.rech}, \code{inciden.rech}, and \code{pumping.rech}.
  The well configuration data are described in the \sQuote{Value} section of the
  \code{\link{GetWellConfig}} function.
}
\description{
This function determines the specified-flow boundary conditions for the groundwater-flow model.
These boundary conditions include:
  (1) natural and incidental groundwater recharge at the water table,
  (2) groundwater pumping at production wells, and
  (3) groundwater underflow in the major tributary valleys.
Specified-flow values are saved to disk by rewriting the
\href{https://water.usgs.gov/ogw/modflow/}{MODFLOW} Well Package file (\file{.wel}).
Note that this function is executed after each iteration of \href{http://www.pesthomepage.org/}{PEST}.
}
\details{
Files read during execution, and located within the \code{dir.run} directory,
  inlcude the MODFLOW hydraulic conductivity reference files \file{hk1.ref},
  \file{hk2.ref}, and \file{hk3.ref} corresponding to model layers 1, 2, and 3, respectively.
  Hydraulic conductivity values are read from a two-dimensional array in
  matrix format with \sQuote{white-space} delimited fields.
  And a binary data file \file{model.rda} containing the following serialized \R objects:
  \code{rs}, \code{misc}, \code{trib}, \code{tr.stress.periods}, and \code{ss.stress.periods}.

  \code{rs} is an object of RasterStack class with raster layers \dQuote{lay1.top},
  \dQuote{lay1.bot}, \dQuote{lay2.bot}, and \dQuote{lay3.bot}.
  These raster layers describe the geometry of the model grid; that is,
  the upper and lower elevation of model layer 1, and the bottom elevations of model layers 2 and 3.
  Missing cell values (equal to NA) indicate inactive model cells lying outside of the model domain.

  \code{misc} is a data.frame object with miscellaneous seepage,
  such as from the \sQuote{Bellevue Waste Water Treatment Plant ponds} and the \sQuote{Bypass Canal}.
  This object is comprised of the following components:
  \bold{lay}, \bold{row}, \bold{col} are integer values specifying a
  model cell's layer, row, and column index, respectively; and
  \bold{ss}, \bold{199501}, \bold{199502}, \dots, \bold{201012} are numeric values of
  elevation during each stress period, respectively,
  in meters above the North American Vertical Datum of 1988.

  \code{trib} is a data.frame object with default values for the
  long-term mean underflows in each of the tributary basins.
  The object is comprised of the following components:
  \bold{Name} is a unique identifier for the tributary basin;
  \bold{lay}, \bold{row}, \bold{col} are \code{integer} values of a
  model cell's layer, row, and column index, respectively; and
  \bold{ss}, \bold{199501}, \bold{199502}, \dots, \bold{201012} are numeric values of
  underflow during each stress period, respectively, in cubic meters per day.

  \code{tr.stress.periods} is a vector of Date values giving the start and end dates for
  stress periods in the model simulation period (1995--2010).

  \code{ss.stress.periods} is a vector of Date values giving the start and end dates for
  stress periods used to define steady-state conditions.

  \code{reduction} is a numeric default value for the signal amplitude reduction algorithm,
  a dimensionless quantity.

  \code{d.in.mv.ave} is a numeric default value for the number of days in the
  moving average subset.
}
\examples{
\dontrun{
  dir.run <- file.path(getwd(), "model/model1")
  UpdateWaterBudget(dir.run, "wrv_mfusg", qa.tables = "si",
                    ss.interval = c("1998-01-01", "2011-01-01"))
}

}
\seealso{
\code{\link{RunWaterBalance}}, \code{\link{GetSeasonalMult}}
}
\author{
J.C. Fisher, U.S. Geological Survey, Idaho Water Science Center
}
\keyword{utilities}
